requirements.txt
Django==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1

